Move utils, improve evid logic
